fix: type hint was wrong type in dx data generator
Better type hinting support means this was now actually triggering an error. However, it is currently returning numpy types, which is a bug in core. Initial random.seed(seed) now prints this error, and is expected to be fixed in next release.

> DeprecationWarning: Seeding based on hashing is deprecated
since Python 3.9 and will be removed in a subsequent version. The only
supported seed types are: None, int, float, str, bytes, and bytearray.

See: deephaven/deephaven-core#4933