fix: type hint was wrong type in dx data generator (#155)
Better type hinting support means this mistake was now actually
triggering an error.

Engine is currently returning numpy types instead of python types, which
is a bug in core. Initial random.seed(seed) is printing this error
because of the numpy.int64 type, but is expected to be fixed in the next
dhc release.

> DeprecationWarning: Seeding based on hashing is deprecated
since Python 3.9 and will be removed in a subsequent version. The only
supported seed types are: None, int, float, str, bytes, and bytearray.

See: deephaven/deephaven-core#4933