fix: Hooks accepting liveness objects and functions should manage their lifetime (#258)
Previously, RenderContext managed a single liveness scope, kept open
while invoking the component's function. This missed a variety of cases,
which this commit aims to correct, with a few strategies. RenderContext
itself is modified to track a collection of LivenessScopes, and after
each render will release those scopes which are no longer relevant.
Contributors to that collection can call `manage()` to re-submit a scope
that was used in the previous render, but older scopes that were not
reused cannot be submitted in this way - they should be re-created from
scratch. If the render fails, old and new liveness scopes will be
retained, as a way to protect against a subsequent successful render
from reusing a scope that might not have had a chance to be managed. If
render succeeds, only new and reused scopes will be retained.

First, there is code that may run outside of the component's own
rendering, such as in a table listener, or an event handler. This we
leave to users to decide - a `use_liveness_scope` hook is available to
wrap functions that will be used externally. Wrapping functions in this
way allows the hook to know that any accumulated objects should be kept
alive only for the very next render call.

Next, `use_state`, `use_effect`, and `use_memo` will wrap any functions
that they accept in a liveness scope, and ensure that the render context
will manage them as long as the values or dependencies do not change.

---------

Co-authored-by: Joe <josephnumainville@deephaven.io>
Co-authored-by: Mike Bender <mikebender@deephaven.io>