fix: Call listener when do_replay is True for use_table_listener_hook (#313)
Fixes #290

Theoretically, there is some danger in this implementation, as if
dependencies change the listener will be called again, but I think it
makes more sense than just checking the first time, as the dependencies
changing means a completely different call. Maybe we need more danger
signs as has been alluded to. Maybe that's a docs issue.