feat: Add utilities package (#331)
Adds new utilities and packaging packages. These can be used for
cross-plugin functionality. The first versions have functions that make
it simpler to register js plugins.

Part of the work for #316 and #194 

After this is merged I will upload the initial versions to pypi, then
use that for the follow-up PR to fix the issues themselves.