docs: list_view spec (#352)
Adds spec for `list_view`

There is one major additions on top of what has been discussed in the
picker spec.

I see it as an oversight (or at least not ideal) to some extent that
there's no simple way for embedded buttons to access their parent's keys
easily, as far as I can tell. I suppose it's simple enough to bind item
keys to button events if you're doing this in React, but I think we'll
want a way to do that easily, especially with tables, because it doesn't
make sense to pass a button for each item in a ticking table. The least
intrusive method here is returning extra args for the item key and
section key that a button is embedded in that can be ignored (using the
changes implemented in #271).