fix: Memoize use_table_data listener (#428)
Because the listener function was not memoized in `use_table_data`, the
`use_effect` within `use_table_listener` was getting called
unnecessarily, which means in some cases table updates can be missed
during the start and stop of the listener.

Added fix and caution message in `use_table_listener` for now.

It might be worthwhile making `use_table_listener` more robust in cases
where arguments changed, but the table specifically did not. This is not
trivial though.