fix: gapminder to be compatible with Pandas 2.0.3 (#586)
The current version of the ticking Gapminder dataset uses a feature only
available in Pandas 2.1.4. Deephaven will work with Python 3.8, but
Pandas 2.1 has a minimum Python version of 3.9. So, the feature that
Gapminder depends on is not guaranteed to be available to DH. This PR
removes that feature and replicates the functionality with an older
feature.