feat: Make ui.panel align-items start by default
ui.panels implicitly wrap things in a flex, with a direction set to column. Column makes sense, as tables typically are wide, and stacking them makes more sense by default that side by side.

However, a key distinction is with row, the horizonal axis alignment is equivalent to start. In column the horizonal axis (which is now a cross-axis) is stretch.

I think we should set align-items: start for the implicit ui.flex column wrapper of ui.panel so that it doesn't stretch buttons in the horizontal cross-axis by default.