ci: Add workflow dispatch to Build Docs (#773)
Fixes #724 

Have tested on my own fork. Adds a workflow dispatch to
`build-main-docs.yml`. It's in a slightly funky spot because release
docs are automatically synced separately in a different workflow, but
this could still be used to sync docs from a specific release, without
actually releasing the package itself.