fix: Fix widget plugin python warnings (#740)
Fixes #713 
Mostly minor warnings fixed
The main fix of consequence is a new class, `CommonJsPlugin`, in
utilities. This should be inherited from by all plugins that use
`JsPlugin` because `create_js_plugin` actually requires something of the
form `CommonJsPlugin` as `JsPlugin` doesn't specify the inputs it takes.
I made `CommonJsPlugin` the default but a `JsPlugin` can still be passed
so that this doesn't break existing plugins. We should do this first and
release a new `utilities` package. I will make a follow-up PR that will
fix the inheritance on all our packages to ensure the typing is correct.