feat: UI Dialog and DialogTrigger Components (#953)
closes https://github.com/deephaven/deephaven-plugins/issues/933
closes https://github.com/deephaven/deephaven-plugins/issues/863

Note that there is a bug in the spectrum theme that causes the underlay
to render as transparent. It is fixed here:
https://github.com/deephaven/web-client-ui/pull/2267
The underlay may render incorrectly depending on your version of
`web-client-ui`
Because of this bug, we are waiting for a new version before adding a
render test for dialog. See
https://github.com/deephaven/deephaven-plugins/issues/952

---------

Co-authored-by: Brian Ingles <github@emeraldwalk.com>
Co-authored-by: margaretkennedy <82049573+margaretkennedy@users.noreply.github.com>
Co-authored-by: Joe <josephnumainville@deephaven.io>