feat: Show loading spinners immediately in ui (#1023)
I can't find any tickets on this even though I thought we had some.

Fixes a few issues
1. Immediately open a panel and show a loading spinner before server
rendering is complete The panel will be re-used and have its title
updated if necessary when rendering completes. Works with multiple
panels (opens 1 panel for loading and then re-uses for the 1st of
multiple panels)
2. Shows a loader for all panels on page reload until rendering is
complete. Previously, we rendered a blank panel until re-hydration was
complete
3. If a widget throws a rendering error, the state is reset to loading
when the "Reload" button is pressed. A loader is immediately shown until
the reload is over instead of just showing the error message.