feat: `dx.indicator` spec (#1062)
spec for #1019 

Add `dx.indicator` with `NotImplementedError`.

Instead of just enabling one indicator, you can create multiple at one
time with a plot by that makes subplots of indicators. This enables, for
example, the creation of a new indicator automatically as a new symbol
ticks in.
The default behavior is only a `number` if you only provide `value` and
`number+delta` if you provide `reference` as well.
https://plotly.com/python/indicator/

Syntax is otherwise generally in line with other `dx` functions and most
other behavior is fairly straightforward.