ci: Publish versioned docs fix (#1129)
Reverting the switch to 2 actions and just removing the branch filter
that was causing preview docs not to run on PRs. Since the branch filter
doesn't seem to work, doesn't seem like much benefit to have 2 separate
actions that are nearly identical for now.

The `Test Modified Plugin` workflow filters the push branches and tags
allowed. If we want to keep separate actions, I confirmed in my fork
that `branches: [main, v*]` included events on the tag `v0.1`. So we
could include the tags as "branches" for the prod docs workflow.