chore: DOC-714 Add local docs preview (#1163)
Adds `npm run docs` which will run docs from source code. This does not
generate the API param tables, but shows the placeholder value that the
sphinx plugin replaces.

I initially had something to run `plugin_builder.py` in watch mode and
the docs docker image, but `run-p` would cause this to prevent shutdown
with `ctrl+c` for some reason. And it was significantly slower because
it would need to rebuild the Python plugin