build: DH-19353: Add transpiled files and types to plotly-express build (#1170)
Just adding back for plotly-express since we weren't publishing types or
non-bundled code for the other packages. If we need them in the future
we can add them at that time. I didn't want to potentially bork adding
them or spend time testing everything is configured properly for other
packages if we just need to add the build step back for plotly-express.