ci: Update e2e snapshots (#1225)
These were updated as part of another PR on the main branch, so needed
to update these snapshots for the v0.31 release branch as well in case
we need to pick more back to this branch

The update accounts for the now correct coloring of numbers in aggs. Also pins plotly version so we don't have issues with pre-commit.