chore: Increase default concurrency for plugin dev (#1270)
The default concurrency value for `lerna` is to use the number of
logical cores. In cases where the number of logical cores is < the
number of plugins, `npm start:packages` will only build some of the
plugins.

For example, on my Mac, I have 10 logical cores, and we now have 11
plugins. This resulted in `ui` plugin not being built.

This PR increases the default to 20 which gives us some cushion to grow
in the number of plugins.

## Testing
I have tested this by running `npm start` on my machine which has 10
logical cores. All 11 plugins ran in watch mode. Without the
`--concurrency 20` flag, only 10 of them run.